+++ tests/testthat/test-hooks.R
@@ -283,7 +283,7 @@ run_test("roxygenize",
     "DESCRIPTION" = test_path("in/DESCRIPTION-no-deps.dcf")
   ),
   file_transformer = function(files) {
-    git2r::init()
+    git_init()
     git2r::add(path = files)
     files
   }
@@ -303,7 +303,7 @@ run_test("readme-rmd-rendered",
       content_2 <- readLines(files[2])
       Sys.sleep(2)
       writeLines(content_2, files[2])
-      git2r::init()
+      git_init()
       git2r::add(path = files)
     }
     files
@@ -321,7 +321,7 @@ run_test("readme-rmd-rendered",
       content_2 <- readLines(files[2])
       Sys.sleep(2)
       writeLines(content_2, files[2])
-      git2r::init()
+      git_init()
       git2r::add(path = files)
     }
     files
@@ -341,7 +341,7 @@ run_test("readme-rmd-rendered",
       content_2 <- readLines(files[2])
       Sys.sleep(2)
       writeLines(content_2, files[2])
-      git2r::init()
+      git_init()
       git2r::add(path = files[1])
     }
     files
@@ -355,7 +355,7 @@ run_test("readme-rmd-rendered",
   error_msg = NULL,
   msg = NULL,
   file_transformer = function(files) {
-    git2r::init()
+    git_init()
     git2r::add(path = files[1])
     files
   }
@@ -368,7 +368,7 @@ run_test("readme-rmd-rendered",
   error_msg = NULL,
   msg = NULL,
   file_transformer = function(files) {
-    git2r::init()
+    git_init()
     git2r::add(path = files[1])
     files
   }